Move qtviewer to _qt folder
